Fix config parsing and added request check
ConfigParser is the python2 version of the import, whereas configparser
is the python3 version. The syntax is a lot simpler too, and does not
require the io import.

If the id did not point to a torrent, r.json()['status'] would return
with 'failure' and would not have a ['response'] section. A check for
['success'] prevents the script from failing when the id doesn't point
to anything.

Added time.sleep(2) to stay within the request limit of the GGn API.