Add snatch.py and relative sample config options
snatch.py is used for actually downloading the .torrent files from the
site. The script is currently hardcoded in the number of torrents it
downloads, but will read in either a configuration option or a command
argument in the future.

Added the authkey and torrent_pass options to the sampel configuration
file for adding the relavent info for downloading torrent files.

Added emacs files and python cache directory to .gitignore for editor
support. Added 'failed' to .gitignore, but haven't implemented actually
outputting failed requests to the file yet.

Increased timeout time of api calls to 2s instead of 1s.